extracting global functions and effects
