+++ src/App.tsx
@@ -368,6 +368,7 @@ function App() {
     isLightTheme,
     visualAdjustmentsKey,
     pinnedFolders,
+    isAnyModalOpen,
   } = useAppState();
 
   const { sortedImageList } = useSortedImageList();
@@ -410,417 +411,43 @@ function App() {
     handleSelectSubfolder,
     handleLibraryRefresh,
     handleToggleFolder,
+    handleToggleFullScreen,
+    handleCopyAdjustments,
+    handlePasteAdjustments,
+    handleAutoAdjustments,
+    handleRate,
+    handleSetColorLabel,
+    getCommonTags,
+    handleTagsChanged,
+    closeConfirmModal,
+    handlePasteFiles,
+    requestFullResolution,
+    handleFullResolutionLogic,
+    handleZoomChange,
+    handleUserTransform,
+    handleSavePanorama,
+    handleSaveHdr,
+    handleApplyDenoise,
+    handleSaveDenoisedImage,
+    handleSaveCollage,
+    handleOpenFolder,
+    handleContinueSession,
+    handleGoHome,
+    handleMultiSelectClick,
+    handleLibraryImageSingleClick,
+    handleImageClick,
+    handleClearSelection,
+    handleRenameFiles,
+    handleSaveRename,
+    handleStartImport,
+    handleResetAdjustments,
+    handleImportClick,
   } = useHandlers();
 
   useGlobalEffects();
 
   const { showContextMenu } = useContextMenu();
   const { loading: isThumbnailsLoading } = useThumbnails(imageList, setThumbnails);
-
-  const handleToggleFullScreen = useCallback(() => {
-    if (isFullScreen) {
-      setIsFullScreen(false);
-    } else {
-      if (!selectedImage) {
-        return;
-      }
-      setIsFullScreen(true);
-    }
-  }, [isFullScreen, selectedImage]);
-
-  const handleCopyAdjustments = useCallback(() => {
-    const sourceAdjustments = selectedImage ? adjustments : libraryActiveAdjustments;
-    const adjustmentsToCopy: any = {};
-    for (const key of COPYABLE_ADJUSTMENT_KEYS) {
-      if (Object.prototype.hasOwnProperty.call(sourceAdjustments, key)) {
-        adjustmentsToCopy[key] = sourceAdjustments[key];
-      }
-    }
-    setCopiedAdjustments(adjustmentsToCopy);
-    setIsCopied(true);
-  }, [selectedImage, adjustments, libraryActiveAdjustments]);
-
-  const handlePasteAdjustments = useCallback(
-    (paths?: Array<string>) => {
-      if (!copiedAdjustments || !appSettings) {
-        return;
-      }
-
-      const { mode, includedAdjustments } = appSettings.copyPasteSettings;
-
-      const adjustmentsToApply: Partial<Adjustments> = {};
-
-      for (const key of includedAdjustments) {
-        if (Object.prototype.hasOwnProperty.call(copiedAdjustments, key)) {
-          const value = copiedAdjustments[key as keyof Adjustments];
-
-          if (mode === PasteMode.Merge) {
-            const defaultValue = INITIAL_ADJUSTMENTS[key as keyof Adjustments];
-            if (JSON.stringify(value) !== JSON.stringify(defaultValue)) {
-              adjustmentsToApply[key as keyof Adjustments] = value;
-            }
-          } else {
-            adjustmentsToApply[key as keyof Adjustments] = value;
-          }
-        }
-      }
-
-      if (Object.keys(adjustmentsToApply).length === 0) {
-        setIsPasted(true);
-        return;
-      }
-
-      const pathsToUpdate =
-        paths || (multiSelectedPaths.length > 0 ? multiSelectedPaths : selectedImage ? [selectedImage.path] : []);
-      if (pathsToUpdate.length === 0) {
-        return;
-      }
-
-      if (selectedImage && pathsToUpdate.includes(selectedImage.path)) {
-        const newAdjustments = { ...adjustments, ...adjustmentsToApply };
-        setAdjustments(newAdjustments);
-      }
-
-      invoke(Invokes.ApplyAdjustmentsToPaths, { paths: pathsToUpdate, adjustments: adjustmentsToApply }).catch(
-        (err) => {
-          console.error('Failed to paste adjustments to multiple images:', err);
-          setError(`Failed to paste adjustments: ${err}`);
-        },
-      );
-      setIsPasted(true);
-    },
-    [copiedAdjustments, appSettings, multiSelectedPaths, selectedImage, adjustments, setAdjustments],
-  );
-
-  const handleAutoAdjustments = async () => {
-    if (!selectedImage) {
-      return;
-    }
-    try {
-      const autoAdjustments: Adjustments = await invoke(Invokes.CalculateAutoAdjustments);
-      setAdjustments((prev: Adjustments) => {
-        const newAdjustments = { ...prev, ...autoAdjustments };
-        newAdjustments.sectionVisibility = {
-          ...prev.sectionVisibility,
-          ...autoAdjustments.sectionVisibility,
-        };
-
-        return newAdjustments;
-      });
-    } catch (err) {
-      console.error('Failed to calculate auto adjustments:', err);
-      setError(`Failed to apply auto adjustments: ${err}`);
-    }
-  };
-
-  const handleRate = useCallback(
-    (newRating: number, paths?: Array<string>) => {
-      const pathsToRate =
-        paths || (multiSelectedPaths.length > 0 ? multiSelectedPaths : selectedImage ? [selectedImage.path] : []);
-      if (pathsToRate.length === 0) {
-        return;
-      }
-
-      let currentRating = 0;
-      if (selectedImage && pathsToRate.includes(selectedImage.path)) {
-        currentRating = adjustments.rating;
-      } else if (libraryActivePath && pathsToRate.includes(libraryActivePath)) {
-        currentRating = libraryActiveAdjustments.rating;
-      }
-
-      const finalRating = newRating === currentRating ? 0 : newRating;
-
-      setImageRatings((prev: Record<string, number>) => {
-        const newRatings = { ...prev };
-        pathsToRate.forEach((path: string) => {
-          newRatings[path] = finalRating;
-        });
-        return newRatings;
-      });
-
-      if (selectedImage && pathsToRate.includes(selectedImage.path)) {
-        setAdjustments((prev: Adjustments) => ({ ...prev, rating: finalRating }));
-      }
-
-      if (libraryActivePath && pathsToRate.includes(libraryActivePath)) {
-        setLibraryActiveAdjustments((prev) => ({ ...prev, rating: finalRating }));
-      }
-
-      invoke(Invokes.ApplyAdjustmentsToPaths, { paths: pathsToRate, adjustments: { rating: finalRating } }).catch(
-        (err) => {
-          console.error('Failed to apply rating to paths:', err);
-          setError(`Failed to apply rating: ${err}`);
-        },
-      );
-    },
-    [
-      multiSelectedPaths,
-      selectedImage,
-      libraryActivePath,
-      adjustments.rating,
-      libraryActiveAdjustments.rating,
-      setAdjustments,
-    ],
-  );
-
-  const handleSetColorLabel = useCallback(
-    async (color: string | null, paths?: Array<string>) => {
-      const pathsToUpdate =
-        paths || (multiSelectedPaths.length > 0 ? multiSelectedPaths : selectedImage ? [selectedImage.path] : []);
-      if (pathsToUpdate.length === 0) {
-        return;
-      }
-      const primaryPath = selectedImage?.path || libraryActivePath;
-      const primaryImage = imageList.find((img: ImageFile) => img.path === primaryPath);
-      let currentColor = null;
-      if (primaryImage && primaryImage.tags) {
-        const colorTag = primaryImage.tags.find((tag: string) => tag.startsWith('color:'));
-        if (colorTag) {
-          currentColor = colorTag.substring(6);
-        }
-      }
-      const finalColor = color !== null && color === currentColor ? null : color;
-      try {
-        await invoke(Invokes.SetColorLabelForPaths, { paths: pathsToUpdate, color: finalColor });
-
-        setImageList((prevList: Array<ImageFile>) =>
-          prevList.map((image: ImageFile) => {
-            if (pathsToUpdate.includes(image.path)) {
-              const otherTags = (image.tags || []).filter((tag: string) => !tag.startsWith('color:'));
-              const newTags = finalColor ? [...otherTags, `color:${finalColor}`] : otherTags;
-              return { ...image, tags: newTags };
-            }
-            return image;
-          }),
-        );
-      } catch (err) {
-        console.error('Failed to set color label:', err);
-        setError(`Failed to set color label: ${err}`);
-      }
-    },
-    [multiSelectedPaths, selectedImage, libraryActivePath, imageList],
-  );
-
-  const getCommonTags = useCallback(
-    (paths: string[]): { tag: string; isUser: boolean }[] => {
-      if (paths.length === 0) return [];
-      const imageFiles = imageList.filter((img) => paths.includes(img.path));
-      if (imageFiles.length === 0) return [];
-
-      const allTagsSets = imageFiles.map((img) => {
-        const tagsWithPrefix = (img.tags || []).filter((t) => !t.startsWith('color:'));
-        return new Set(tagsWithPrefix);
-      });
-
-      if (allTagsSets.length === 0) return [];
-
-      const commonTagsWithPrefix = allTagsSets.reduce((intersection, currentSet) => {
-        return new Set([...intersection].filter((tag) => currentSet.has(tag)));
-      });
-
-      return Array.from(commonTagsWithPrefix)
-        .map((tag) => ({
-          tag: tag.startsWith('user:') ? tag.substring(5) : tag,
-          isUser: tag.startsWith('user:'),
-        }))
-        .sort((a, b) => a.tag.localeCompare(b.tag));
-    },
-    [imageList],
-  );
-
-  const handleTagsChanged = useCallback(
-    (changedPaths: string[], newTags: { tag: string; isUser: boolean }[]) => {
-      setImageList((prevList) =>
-        prevList.map((image) => {
-          if (changedPaths.includes(image.path)) {
-            const colorTags = (image.tags || []).filter((t) => t.startsWith('color:'));
-            const prefixedNewTags = newTags.map((t) => (t.isUser ? `user:${t.tag}` : t.tag));
-            const finalTags = [...colorTags, ...prefixedNewTags].sort();
-            return { ...image, tags: finalTags.length > 0 ? finalTags : null };
-          }
-          return image;
-        }),
-      );
-    },
-    [setImageList],
-  );
-
-  const closeConfirmModal = () => setConfirmModalState({ ...confirmModalState, isOpen: false });
-
-  const handlePasteFiles = useCallback(
-    async (mode = 'copy') => {
-      if (copiedFilePaths.length === 0 || !currentFolderPath) {
-        return;
-      }
-      try {
-        if (mode === 'copy')
-          await invoke(Invokes.CopyFiles, { sourcePaths: copiedFilePaths, destinationFolder: currentFolderPath });
-        else {
-          await invoke(Invokes.MoveFiles, { sourcePaths: copiedFilePaths, destinationFolder: currentFolderPath });
-          setCopiedFilePaths([]);
-        }
-        await refreshImageList();
-      } catch (err) {
-        setError(`Failed to ${mode} files: ${err}`);
-      }
-    },
-    [copiedFilePaths, currentFolderPath, refreshImageList],
-  );
-
-  const requestFullResolution = useCallback(
-    debounce((currentAdjustments: any, key: string) => {
-      if (!selectedImage?.path) return;
-
-      const jobId = ++previewJobIdRef.current;
-
-      invoke<ArrayBuffer>(Invokes.GenerateFullscreenPreview, {
-        jsAdjustments: currentAdjustments,
-      })
-        .then((buffer) => {
-          if (jobId >= latestRenderedJobIdRef.current) {
-            latestRenderedJobIdRef.current = jobId;
-            const blob = new Blob([buffer], { type: 'image/jpeg' });
-            const url = URL.createObjectURL(blob);
-
-            setFinalPreviewUrl((prevUrl) => {
-              if (prevUrl && prevUrl.startsWith('blob:')) URL.revokeObjectURL(prevUrl);
-              return url;
-            });
-
-            fullResCacheKeyRef.current = key;
-            setIsLoadingFullRes(false);
-          }
-        })
-        .catch((error: any) => {
-          if (jobId >= latestRenderedJobIdRef.current) {
-            console.error('Failed to generate high resolution preview:', error);
-            fullResCacheKeyRef.current = null;
-            setIsLoadingFullRes(false);
-          }
-        });
-    }, 100),
-    [selectedImage?.path],
-  );
-
-  useEffect(() => {
-    if ((isFullScreen || isHighResNeeded) && selectedImage?.isReady) {
-      if (fullResCacheKeyRef.current !== visualAdjustmentsKey) {
-        setIsLoadingFullRes(true);
-        requestFullResolution(adjustments, visualAdjustmentsKey);
-      }
-    } else if (!isFullScreen && !isHighResNeeded) {
-      if (requestFullResolution.cancel) {
-        requestFullResolution.cancel();
-      }
-      setIsLoadingFullRes(false);
-    }
-  }, [adjustments, isFullScreen, isHighResNeeded, selectedImage?.isReady, requestFullResolution, visualAdjustmentsKey]);
-
-  const handleFullResolutionLogic = useCallback(
-    (targetZoomPercent: number) => {
-      if (appSettings?.enableZoomHifi === false) {
-        return;
-      }
-
-      if (!initialFitScale) {
-        return;
-      }
-      const highResThreshold = Math.max(initialFitScale * 2, 0.5);
-      const needsFullRes = targetZoomPercent > highResThreshold;
-      const previewIsAlreadyFullRes = previewSize.width >= originalSize.width;
-
-      if (needsFullRes && !previewIsAlreadyFullRes) {
-        setIsHighResNeeded(true);
-      } else {
-        setIsHighResNeeded(false);
-      }
-    },
-    [initialFitScale, previewSize.width, originalSize.width, appSettings],
-  );
-
-  const handleZoomChange = useCallback(
-    (zoomValue: number, fitToWindow: boolean = false) => {
-      let targetZoomPercent: number;
-      const orientationSteps = adjustments.orientationSteps || 0;
-      const isSwapped = orientationSteps === 1 || orientationSteps === 3;
-      const effectiveOriginalWidth = isSwapped ? originalSize.height : originalSize.width;
-      const effectiveOriginalHeight = isSwapped ? originalSize.width : originalSize.height;
-      if (fitToWindow) {
-        if (
-          effectiveOriginalWidth > 0 &&
-          effectiveOriginalHeight > 0 &&
-          baseRenderSize.width > 0 &&
-          baseRenderSize.height > 0
-        ) {
-          const originalAspect = effectiveOriginalWidth / effectiveOriginalHeight;
-          const baseAspect = baseRenderSize.width / baseRenderSize.height;
-          if (originalAspect > baseAspect) {
-            targetZoomPercent = baseRenderSize.width / effectiveOriginalWidth;
-          } else {
-            targetZoomPercent = baseRenderSize.height / effectiveOriginalHeight;
-          }
-        } else {
-          targetZoomPercent = 1.0;
-        }
-      } else {
-        targetZoomPercent = zoomValue;
-      }
-      targetZoomPercent = Math.max(0.1, Math.min(2.0, targetZoomPercent));
-      let transformZoom = 1.0;
-      if (
-        effectiveOriginalWidth > 0 &&
-        effectiveOriginalHeight > 0 &&
-        baseRenderSize.width > 0 &&
-        baseRenderSize.height > 0
-      ) {
-        const originalAspect = effectiveOriginalWidth / effectiveOriginalHeight;
-        const baseAspect = baseRenderSize.width / baseRenderSize.height;
-        if (originalAspect > baseAspect) {
-          transformZoom = (targetZoomPercent * effectiveOriginalWidth) / baseRenderSize.width;
-        } else {
-          transformZoom = (targetZoomPercent * effectiveOriginalHeight) / baseRenderSize.height;
-        }
-      }
-      isProgrammaticZoom.current = true;
-      setZoom(transformZoom);
-      handleFullResolutionLogic(targetZoomPercent);
-    },
-    [originalSize, baseRenderSize, handleFullResolutionLogic, adjustments.orientationSteps],
-  );
-
-  const handleUserTransform = useCallback(
-    (transformState: TransformState) => {
-      if (isProgrammaticZoom.current) {
-        isProgrammaticZoom.current = false;
-        return;
-      }
-
-      setZoom(transformState.scale);
-
-      if (originalSize.width > 0 && baseRenderSize.width > 0) {
-        const orientationSteps = adjustments.orientationSteps || 0;
-        const isSwapped = orientationSteps === 1 || orientationSteps === 3;
-        const effectiveOriginalWidth = isSwapped ? originalSize.height : originalSize.width;
-
-        const targetZoomPercent = (baseRenderSize.width * transformState.scale) / effectiveOriginalWidth;
-        handleFullResolutionLogic(targetZoomPercent);
-      }
-    },
-    [originalSize, baseRenderSize, handleFullResolutionLogic, adjustments.orientationSteps],
-  );
-
-  const isAnyModalOpen =
-    isCreateFolderModalOpen ||
-    isRenameFolderModalOpen ||
-    isRenameFileModalOpen ||
-    isImportModalOpen ||
-    isCopyPasteSettingsModalOpen ||
-    confirmModalState.isOpen ||
-    panoramaModalState.isOpen ||
-    cullingModalState.isOpen ||
-    collageModalState.isOpen ||
-    denoiseModalState.isOpen ||
-    negativeModalState.isOpen;
 
   useKeyboardShortcuts({
     isModalOpen: isAnyModalOpen,
@@ -869,939 +496,6 @@ function App() {
     baseRenderSize,
     originalSize,
   });
-
-  useEffect(() => {
-    let isEffectActive = true;
-    const listeners = [
-      listen('preview-update-uncropped', (event: any) => {
-        if (isEffectActive) {
-          setUncroppedAdjustedPreviewUrl(event.payload);
-        }
-      }),
-      listen('histogram-update', (event: any) => {
-        if (isEffectActive) {
-          setHistogram(event.payload);
-        }
-      }),
-      listen('open-with-file', (event: any) => {
-        if (isEffectActive) {
-          setInitialFileToOpen(event.payload as string);
-        }
-      }),
-      listen('waveform-update', (event: any) => {
-        if (isEffectActive) {
-          setWaveform(event.payload);
-        }
-      }),
-      listen('thumbnail-generated', (event: any) => {
-        if (isEffectActive) {
-          const { path, data, rating } = event.payload;
-          if (data) {
-            setThumbnails((prev) => ({ ...prev, [path]: data }));
-          }
-          if (rating !== undefined) {
-            setImageRatings((prev) => ({ ...prev, [path]: rating }));
-          }
-        }
-      }),
-      listen('ai-model-download-start', (event: any) => {
-        if (isEffectActive) {
-          setAiModelDownloadStatus(event.payload);
-        }
-      }),
-      listen('ai-model-download-finish', () => {
-        if (isEffectActive) {
-          setAiModelDownloadStatus(null);
-        }
-      }),
-      listen('indexing-started', () => {
-        if (isEffectActive) {
-          setIsIndexing(true);
-          setIndexingProgress({ current: 0, total: 0 });
-        }
-      }),
-      listen('indexing-progress', (event: any) => {
-        if (isEffectActive) {
-          setIndexingProgress(event.payload);
-        }
-      }),
-      listen('indexing-finished', () => {
-        if (isEffectActive) {
-          setIsIndexing(false);
-          setIndexingProgress({ current: 0, total: 0 });
-          if (currentFolderPathRef.current) {
-            const refreshImageList = async () => {
-              try {
-                const list: ImageFile[] = await invoke(Invokes.ListImagesInDir, { path: currentFolderPathRef.current });
-                if (Array.isArray(list)) {
-                  setImageList(list);
-                }
-              } catch (err) {
-                console.error('Failed to refresh after indexing:', err);
-              }
-            };
-            refreshImageList();
-          }
-        }
-      }),
-      listen('batch-export-progress', (event: any) => {
-        if (isEffectActive) {
-          setExportState((prev: ExportState) => ({ ...prev, progress: event.payload }));
-        }
-      }),
-      listen('export-complete', () => {
-        if (isEffectActive) {
-          setExportState((prev: ExportState) => ({ ...prev, status: Status.Success }));
-        }
-      }),
-      listen('export-error', (event) => {
-        if (isEffectActive) {
-          setExportState((prev: ExportState) => ({
-            ...prev,
-            status: Status.Error,
-            errorMessage: typeof event.payload === 'string' ? event.payload : 'An unknown export error occurred.',
-          }));
-        }
-      }),
-      listen('export-cancelled', () => {
-        if (isEffectActive) {
-          setExportState((prev: ExportState) => ({ ...prev, status: Status.Cancelled }));
-        }
-      }),
-      listen('import-start', (event: any) => {
-        if (isEffectActive) {
-          setImportState({
-            errorMessage: '',
-            path: '',
-            progress: { current: 0, total: event.payload.total },
-            status: Status.Importing,
-          });
-        }
-      }),
-      listen('import-progress', (event: any) => {
-        if (isEffectActive) {
-          setImportState((prev: ImportState) => ({
-            ...prev,
-            path: event.payload.path,
-            progress: { current: event.payload.current, total: event.payload.total },
-          }));
-        }
-      }),
-      listen('import-complete', () => {
-        if (isEffectActive) {
-          setImportState((prev: ImportState) => ({ ...prev, status: Status.Success }));
-          refreshAllFolderTrees();
-          if (currentFolderPathRef.current) {
-            handleSelectSubfolder(currentFolderPathRef.current, false);
-          }
-        }
-      }),
-      listen('import-error', (event) => {
-        if (isEffectActive) {
-          setImportState((prev: ImportState) => ({
-            ...prev,
-            errorMessage: typeof event.payload === 'string' ? event.payload : 'An unknown import error occurred.',
-            status: Status.Error,
-          }));
-        }
-      }),
-      listen('denoise-progress', (event: any) => {
-        if (isEffectActive) {
-          setDenoiseModalState((prev) => ({ ...prev, progressMessage: event.payload as string }));
-        }
-      }),
-      listen('denoise-complete', (event: any) => {
-        if (isEffectActive) {
-          const payload = event.payload;
-          const isObject = typeof payload === 'object' && payload !== null;
-
-          setDenoiseModalState((prev) => ({
-            ...prev,
-            isProcessing: false,
-            previewBase64: isObject ? payload.denoised : payload,
-            originalBase64: isObject ? payload.original : null,
-            progressMessage: null,
-          }));
-        }
-      }),
-      listen('denoise-error', (event: any) => {
-        if (isEffectActive) {
-          setDenoiseModalState((prev) => ({
-            ...prev,
-            isProcessing: false,
-            error: String(event.payload),
-            progressMessage: null,
-          }));
-        }
-      }),
-    ];
-    return () => {
-      isEffectActive = false;
-      listeners.forEach((p) => p.then((unlisten) => unlisten()));
-    };
-  }, [refreshAllFolderTrees, handleSelectSubfolder]);
-
-  useEffect(() => {
-    if ([Status.Success, Status.Error, Status.Cancelled].includes(exportState.status)) {
-      const timeoutDuration = exportState.status === Status.Success ? 5000 : 3000;
-
-      const timer = setTimeout(() => {
-        setExportState({ status: Status.Idle, progress: { current: 0, total: 0 }, errorMessage: '' });
-      }, timeoutDuration);
-      return () => clearTimeout(timer);
-    }
-  }, [exportState.status]);
-
-  useEffect(() => {
-    if ([Status.Success, Status.Error].includes(importState.status)) {
-      const timer = setTimeout(() => {
-        setImportState({ status: Status.Idle, progress: { current: 0, total: 0 }, path: '', errorMessage: '' });
-      }, IMPORT_TIMEOUT);
-
-      return () => clearTimeout(timer);
-    }
-  }, [importState.status]);
-
-  useEffect(() => {
-    if (libraryActivePath) {
-      invoke(Invokes.LoadMetadata, { path: libraryActivePath })
-        .then((metadata: any) => {
-          if (metadata.adjustments && !metadata.adjustments.is_null) {
-            const normalized: Adjustments = normalizeLoadedAdjustments(metadata.adjustments);
-            setLibraryActiveAdjustments(normalized);
-          } else {
-            setLibraryActiveAdjustments(INITIAL_ADJUSTMENTS);
-          }
-        })
-        .catch((err) => {
-          console.error('Failed to load metadata for library active image', err);
-          setLibraryActiveAdjustments(INITIAL_ADJUSTMENTS);
-        });
-    } else {
-      setLibraryActiveAdjustments(INITIAL_ADJUSTMENTS);
-    }
-  }, [libraryActivePath]);
-
-  useEffect(() => {
-    let isEffectActive = true;
-
-    const unlistenProgress = listen('panorama-progress', (event: any) => {
-      if (isEffectActive) {
-        setPanoramaModalState((prev: PanoramaModalState) => ({
-          ...prev,
-          error: null,
-          finalImageBase64: null,
-          isOpen: true,
-          progressMessage: event.payload,
-        }));
-      }
-    });
-
-    const unlistenComplete = listen('panorama-complete', (event: any) => {
-      if (isEffectActive) {
-        const { base64 } = event.payload;
-        setPanoramaModalState((prev: PanoramaModalState) => ({
-          ...prev,
-          error: null,
-          finalImageBase64: base64,
-          progressMessage: 'Panorama Ready',
-        }));
-      }
-    });
-
-    const unlistenError = listen('panorama-error', (event: any) => {
-      if (isEffectActive) {
-        setPanoramaModalState((prev: PanoramaModalState) => ({
-          ...prev,
-          error: String(event.payload),
-          finalImageBase64: null,
-          progressMessage: 'An error occurred.',
-        }));
-      }
-    });
-
-    return () => {
-      isEffectActive = false;
-      unlistenProgress.then((f: any) => f());
-      unlistenComplete.then((f: any) => f());
-      unlistenError.then((f: any) => f());
-    };
-  }, []);
-
-  useEffect(() => {
-    let isEffectActive = true;
-
-    const unlistenProgress = listen('hdr-progress', (event: any) => {
-      if (isEffectActive) {
-        setHdrModalState((prev: HdrModalState) => ({
-          ...prev,
-          error: null,
-          finalImageBase64: null,
-          isOpen: true,
-          progressMessage: event.payload,
-        }));
-      }
-    });
-
-    const unlistenComplete = listen('hdr-complete', (event: any) => {
-      if (isEffectActive) {
-        const { base64 } = event.payload;
-        setHdrModalState((prev: HdrModalState) => ({
-          ...prev,
-          error: null,
-          finalImageBase64: base64,
-          progressMessage: 'Hdr Ready',
-        }));
-      }
-    });
-
-    const unlistenError = listen('hdr-error', (event: any) => {
-      if (isEffectActive) {
-        setHdrModalState((prev: HdrModalState) => ({
-          ...prev,
-          error: String(event.payload),
-          finalImageBase64: null,
-          progressMessage: 'An error occurred.',
-        }));
-      }
-    });
-
-    return () => {
-      isEffectActive = false;
-      unlistenProgress.then((f: any) => f());
-      unlistenComplete.then((f: any) => f());
-      unlistenError.then((f: any) => f());
-    };
-  }, []);
-
-  useEffect(() => {
-    let isEffectActive = true;
-
-    const unlistenStart = listen('culling-start', (event: any) => {
-      if (isEffectActive) {
-        setCullingModalState({
-          isOpen: true,
-          progress: { current: 0, total: event.payload, stage: 'Initializing...' },
-          suggestions: null,
-          error: null,
-        });
-      }
-    });
-
-    const unlistenProgress = listen('culling-progress', (event: any) => {
-      if (isEffectActive) {
-        setCullingModalState((prev) => ({ ...prev, progress: event.payload }));
-      }
-    });
-
-    const unlistenComplete = listen('culling-complete', (event: any) => {
-      if (isEffectActive) {
-        setCullingModalState((prev) => ({ ...prev, progress: null, suggestions: event.payload }));
-      }
-    });
-
-    const unlistenError = listen('culling-error', (event: any) => {
-      if (isEffectActive) {
-        setCullingModalState((prev) => ({ ...prev, progress: null, error: String(event.payload) }));
-      }
-    });
-
-    return () => {
-      isEffectActive = false;
-      unlistenStart.then((f) => f());
-      unlistenProgress.then((f) => f());
-      unlistenComplete.then((f) => f());
-      unlistenError.then((f) => f());
-    };
-  }, []);
-
-  const handleSavePanorama = async (): Promise<string> => {
-    if (panoramaModalState.stitchingSourcePaths.length === 0) {
-      const err = 'Source paths for panorama not found.';
-      setPanoramaModalState((prev: PanoramaModalState) => ({ ...prev, error: err }));
-      throw new Error(err);
-    }
-
-    try {
-      const savedPath: string = await invoke(Invokes.SavePanorama, {
-        firstPathStr: panoramaModalState.stitchingSourcePaths[0],
-      });
-      await refreshImageList();
-      return savedPath;
-    } catch (err) {
-      console.error('Failed to save panorama:', err);
-      setPanoramaModalState((prev: PanoramaModalState) => ({ ...prev, error: String(err) }));
-      throw err;
-    }
-  };
-
-  const handleSaveHdr = async (): Promise<string> => {
-    if (hdrModalState.stitchingSourcePaths.length === 0) {
-      const err = 'Source paths for HDR not found.';
-      setHdrModalState((prev: HdrModalState) => ({ ...prev, error: err }));
-      throw new Error(err);
-    }
-
-    try {
-      const savedPath: string = await invoke(Invokes.SaveHdr, {
-        firstPathStr: hdrModalState.stitchingSourcePaths[0],
-      });
-      await refreshImageList();
-      return savedPath;
-    } catch (err) {
-      console.error('Failed to save HDR image:', err);
-      setHdrModalState((prev: HdrModalState) => ({ ...prev, error: String(err) }));
-      throw err;
-    }
-  };
-
-  const handleApplyDenoise = useCallback(
-    async (intensity: number) => {
-      if (!denoiseModalState.targetPath) return;
-
-      setDenoiseModalState((prev) => ({
-        ...prev,
-        isProcessing: true,
-        error: null,
-        progressMessage: 'Starting engine...',
-      }));
-
-      try {
-        await invoke(Invokes.ApplyDenoising, {
-          path: denoiseModalState.targetPath,
-          intensity: intensity,
-        });
-      } catch (err) {
-        setDenoiseModalState((prev) => ({
-          ...prev,
-          isProcessing: false,
-          error: String(err),
-        }));
-      }
-    },
-    [denoiseModalState.targetPath],
-  );
-
-  const handleSaveDenoisedImage = async (): Promise<string> => {
-    if (!denoiseModalState.targetPath) throw new Error('No target path');
-    const savedPath = await invoke<string>(Invokes.SaveDenoisedImage, {
-      originalPathStr: denoiseModalState.targetPath,
-    });
-    await refreshImageList();
-    return savedPath;
-  };
-
-  const handleSaveCollage = async (base64Data: string, firstPath: string): Promise<string> => {
-    try {
-      const savedPath: string = await invoke(Invokes.SaveCollage, {
-        base64Data,
-        firstPathStr: firstPath,
-      });
-      await refreshImageList();
-      return savedPath;
-    } catch (err) {
-      console.error('Failed to save collage:', err);
-      setError(`Failed to save collage: ${err}`);
-      throw err;
-    }
-  };
-
-  useEffect(() => {
-    if (!selectedImage?.isReady) return;
-
-    if (dragIdleTimer.current) {
-      clearTimeout(dragIdleTimer.current);
-    }
-
-    if (isSliderDragging) {
-      if (appSettings?.enableLivePreviews !== false) {
-        applyAdjustments(adjustments, true);
-      }
-
-      dragIdleTimer.current = setTimeout(() => {
-        applyAdjustments(adjustments, false);
-      }, 150);
-    } else {
-      applyAdjustments(adjustments, false);
-      debouncedSave(selectedImage.path, adjustments);
-    }
-
-    return () => {
-      if (dragIdleTimer.current) clearTimeout(dragIdleTimer.current);
-    };
-  }, [
-    adjustments,
-    selectedImage?.path,
-    selectedImage?.isReady,
-    isSliderDragging,
-    applyAdjustments,
-    debouncedSave,
-    appSettings?.enableLivePreviews,
-  ]);
-
-  const handleOpenFolder = async () => {
-    try {
-      const selected = await open({ directory: true, multiple: false, defaultPath: await homeDir() });
-      if (typeof selected === 'string') {
-        setRootPath(selected);
-        await handleSelectSubfolder(selected, true);
-      }
-    } catch (err) {
-      console.error('Failed to open directory dialog:', err);
-      setError('Failed to open folder selection dialog.');
-    }
-  };
-
-  useEffect(() => {
-    if (!rootPath) {
-      setIsLayoutReady(false);
-      return;
-    }
-
-    const timer = setTimeout(() => {
-      setIsLayoutReady(true);
-    }, 100);
-
-    return () => clearTimeout(timer);
-  }, [rootPath]);
-
-  const handleContinueSession = () => {
-    const restore = async () => {
-      if (!appSettings?.lastRootPath) {
-        return;
-      }
-
-      const root = appSettings.lastRootPath;
-      const folderState = appSettings.lastFolderState;
-      const pathToSelect = folderState?.currentFolderPath || root;
-
-      setRootPath(root);
-
-      if (folderState?.expandedFolders) {
-        const newExpandedFolders = new Set(folderState.expandedFolders);
-        newExpandedFolders.add(root);
-        setExpandedFolders(newExpandedFolders);
-      } else {
-        setExpandedFolders(new Set([root]));
-      }
-
-      setIsTreeLoading(true);
-      try {
-        let treeData;
-        if (preloadedDataRef.current.rootPath === root && preloadedDataRef.current.tree) {
-          treeData = await preloadedDataRef.current.tree;
-          console.log('Preload cache hit for folder tree.');
-        } else {
-          treeData = await invoke(Invokes.GetFolderTree, { path: root });
-        }
-        setFolderTree(treeData);
-      } catch (err) {
-        console.error('Failed to restore folder tree:', err);
-      } finally {
-        setIsTreeLoading(false);
-      }
-
-      let preloadedImages: ImageFile[] | undefined = undefined;
-      if (preloadedDataRef.current.currentPath === pathToSelect && preloadedDataRef.current.images) {
-        try {
-          preloadedImages = await preloadedDataRef.current.images;
-          console.log('Preload cache hit for image list.');
-        } catch (e) {
-          console.error('Failed to retrieve preloaded images', e);
-        }
-      }
-
-      await handleSelectSubfolder(pathToSelect, false, preloadedImages);
-    };
-    restore().catch((err) => {
-      console.error('Failed to restore session, folder might be missing:', err);
-      setError('Failed to restore session. The last used folder may have been moved or deleted.');
-      if (appSettings) {
-        handleSettingsChange({ ...appSettings, lastRootPath: null, lastFolderState: null });
-      }
-      handleGoHome();
-      setIsTreeLoading(false);
-    });
-  };
-
-  useEffect(() => {
-    if (!initialFileToOpen || !appSettings) {
-      return;
-    }
-    const parentDir = getParentDir(initialFileToOpen);
-    if (currentFolderPath !== parentDir) {
-      setRootPath(parentDir);
-      handleSelectSubfolder(parentDir, true);
-      return;
-    }
-    const isImageInList = imageList.some((image) => image.path === initialFileToOpen);
-    if (isImageInList) {
-      handleImageSelect(initialFileToOpen);
-      setInitialFileToOpen(null);
-    } else if (!isViewLoading) {
-      console.warn(`'open-with-file' target ${initialFileToOpen} not found in its directory after loading. Aborting.`);
-      setInitialFileToOpen(null);
-    }
-  }, [
-    initialFileToOpen,
-    appSettings,
-    currentFolderPath,
-    imageList,
-    isViewLoading,
-    handleSelectSubfolder,
-    handleImageSelect,
-  ]);
-
-  const handleGoHome = () => {
-    setRootPath(null);
-    setCurrentFolderPath(null);
-    setImageList([]);
-    setImageRatings({});
-    setFolderTree(null);
-    setMultiSelectedPaths([]);
-    setLibraryActivePath(null);
-    setIsLibraryExportPanelVisible(false);
-    setExpandedFolders(new Set());
-  };
-
-  const handleMultiSelectClick = (path: string, event: any, options: MultiSelectOptions) => {
-    const { ctrlKey, metaKey, shiftKey } = event;
-    const isCtrlPressed = ctrlKey || metaKey;
-    const { shiftAnchor, onSimpleClick, updateLibraryActivePath } = options;
-
-    if (shiftKey && shiftAnchor) {
-      const lastIndex = sortedImageList.findIndex((f) => f.path === shiftAnchor);
-      const currentIndex = sortedImageList.findIndex((f) => f.path === path);
-
-      if (lastIndex !== -1 && currentIndex !== -1) {
-        const start = Math.min(lastIndex, currentIndex);
-        const end = Math.max(lastIndex, currentIndex);
-        const range = sortedImageList.slice(start, end + 1).map((f: ImageFile) => f.path);
-        const baseSelection = isCtrlPressed ? multiSelectedPaths : [shiftAnchor];
-        const newSelection = Array.from(new Set([...baseSelection, ...range]));
-
-        setMultiSelectedPaths(newSelection);
-        if (updateLibraryActivePath) {
-          setLibraryActivePath(path);
-        }
-      }
-    } else if (isCtrlPressed) {
-      const newSelection = new Set(multiSelectedPaths);
-      if (newSelection.has(path)) {
-        newSelection.delete(path);
-      } else {
-        newSelection.add(path);
-      }
-
-      const newSelectionArray = Array.from(newSelection);
-      setMultiSelectedPaths(newSelectionArray);
-
-      if (updateLibraryActivePath) {
-        if (newSelectionArray.includes(path)) {
-          setLibraryActivePath(path);
-        } else if (newSelectionArray.length > 0) {
-          setLibraryActivePath(newSelectionArray[newSelectionArray.length - 1]);
-        } else {
-          setLibraryActivePath(null);
-        }
-      }
-    } else {
-      onSimpleClick(path);
-    }
-  };
-
-  const handleLibraryImageSingleClick = (path: string, event: any) => {
-    handleMultiSelectClick(path, event, {
-      shiftAnchor: libraryActivePath,
-      updateLibraryActivePath: true,
-      onSimpleClick: (p: any) => {
-        setMultiSelectedPaths([p]);
-        setLibraryActivePath(p);
-      },
-    });
-  };
-
-  const handleImageClick = (path: string, event: any) => {
-    const inEditor = !!selectedImage;
-    handleMultiSelectClick(path, event, {
-      shiftAnchor: inEditor ? selectedImage.path : libraryActivePath,
-      updateLibraryActivePath: !inEditor,
-      onSimpleClick: handleImageSelect,
-    });
-  };
-
-  useEffect(() => {
-    const invokeWaveForm = async () => {
-      const waveForm: any = await invoke(Invokes.GenerateWaveform).catch((err) =>
-        console.error('Failed to generate waveform:', err),
-      );
-      if (waveForm) {
-        setWaveform(waveForm);
-      }
-    };
-
-    if (isWaveformVisible && selectedImage?.isReady && !waveform) {
-      invokeWaveForm();
-    }
-  }, [isWaveformVisible, selectedImage?.isReady, waveform]);
-
-  useEffect(() => {
-    if (selectedImage && !selectedImage.isReady && selectedImage.path) {
-      let isEffectActive = true;
-
-      const loadMetadataEarly = async () => {
-        try {
-          const metadata: any = await invoke(Invokes.LoadMetadata, { path: selectedImage.path });
-          if (!isEffectActive) return;
-
-          let initialAdjusts;
-          if (metadata.adjustments && !metadata.adjustments.is_null) {
-            initialAdjusts = normalizeLoadedAdjustments(metadata.adjustments);
-          } else {
-            initialAdjusts = { ...INITIAL_ADJUSTMENTS };
-          }
-
-          setLiveAdjustments(initialAdjusts);
-          resetAdjustmentsHistory(initialAdjusts);
-        } catch (err) {
-          console.error('Failed to load metadata early:', err);
-        }
-      };
-
-      loadMetadataEarly();
-
-      const loadFullImageData = async () => {
-        try {
-          const loadImageResult: any = await invoke(Invokes.LoadImage, { path: selectedImage.path });
-          if (!isEffectActive) {
-            return;
-          }
-
-          const { width, height } = loadImageResult;
-          setOriginalSize({ width, height });
-
-          if (appSettings?.editorPreviewResolution) {
-            const maxSize = appSettings.editorPreviewResolution;
-            const aspectRatio = width / height;
-
-            if (width > height) {
-              const pWidth = Math.min(width, maxSize);
-              const pHeight = Math.round(pWidth / aspectRatio);
-              setPreviewSize({ width: pWidth, height: pHeight });
-            } else {
-              const pHeight = Math.min(height, maxSize);
-              const pWidth = Math.round(pHeight * aspectRatio);
-              setPreviewSize({ width: pWidth, height: pHeight });
-            }
-          } else {
-            setPreviewSize({ width: 0, height: 0 });
-          }
-
-          fullResCacheKeyRef.current = null;
-
-          setSelectedImage((currentSelected: SelectedImage | null) => {
-            if (currentSelected && currentSelected.path === selectedImage.path) {
-              return {
-                ...currentSelected,
-                exif: loadImageResult.exif,
-                height: loadImageResult.height,
-                isRaw: loadImageResult.is_raw,
-                isReady: true,
-                metadata: loadImageResult.metadata,
-                originalUrl: null,
-                width: loadImageResult.width,
-              };
-            }
-            return currentSelected;
-          });
-
-          setLiveAdjustments((prev: Adjustments) => {
-            if (!prev.aspectRatio && !prev.crop) {
-              return { ...prev, aspectRatio: loadImageResult.width / loadImageResult.height };
-            }
-            return prev;
-          });
-        } catch (err) {
-          if (isEffectActive) {
-            console.error('Failed to load image:', err);
-            setError(`Failed to load image: ${err}`);
-            setSelectedImage(null);
-          }
-        } finally {
-          if (isEffectActive) {
-            setIsViewLoading(false);
-          }
-        }
-      };
-      loadFullImageData();
-      return () => {
-        isEffectActive = false;
-      };
-    }
-  }, [selectedImage?.path, selectedImage?.isReady, resetAdjustmentsHistory, appSettings?.editorPreviewResolution]);
-
-  const handleClearSelection = () => {
-    if (selectedImage) {
-      setMultiSelectedPaths([selectedImage.path]);
-    } else {
-      setMultiSelectedPaths([]);
-      setLibraryActivePath(null);
-    }
-  };
-
-  const handleRenameFiles = useCallback(async (paths: Array<string>) => {
-    if (paths && paths.length > 0) {
-      setRenameTargetPaths(paths);
-      setIsRenameFileModalOpen(true);
-    }
-  }, []);
-
-  const handleSaveRename = useCallback(
-    async (nameTemplate: string) => {
-      if (renameTargetPaths.length > 0 && nameTemplate) {
-        try {
-          const newPaths: Array<string> = await invoke(Invokes.RenameFiles, {
-            nameTemplate,
-            paths: renameTargetPaths,
-          });
-
-          await refreshImageList();
-
-          if (selectedImage && renameTargetPaths.includes(selectedImage.path)) {
-            const oldPathIndex = renameTargetPaths.indexOf(selectedImage.path);
-
-            if (newPaths[oldPathIndex]) {
-              handleImageSelect(newPaths[oldPathIndex]);
-            } else {
-              handleBackToLibrary();
-            }
-          }
-
-          if (libraryActivePath && renameTargetPaths.includes(libraryActivePath)) {
-            const oldPathIndex = renameTargetPaths.indexOf(libraryActivePath);
-
-            if (newPaths[oldPathIndex]) {
-              setLibraryActivePath(newPaths[oldPathIndex]);
-            } else {
-              setLibraryActivePath(null);
-            }
-          }
-
-          setMultiSelectedPaths(newPaths);
-        } catch (err) {
-          setError(`Failed to rename files: ${err}`);
-        }
-      }
-
-      setRenameTargetPaths([]);
-    },
-    [renameTargetPaths, refreshImageList, selectedImage, libraryActivePath, handleImageSelect, handleBackToLibrary],
-  );
-
-  const handleStartImport = async (settings: AppSettings) => {
-    if (importSourcePaths.length > 0 && importTargetFolder) {
-      invoke(Invokes.ImportFiles, {
-        destinationFolder: importTargetFolder,
-        settings: settings,
-        sourcePaths: importSourcePaths,
-      }).catch((err) => {
-        console.error('Failed to start import:', err);
-        setImportState({ status: Status.Error, errorMessage: `Failed to start import: ${err}` });
-      });
-    }
-  };
-
-  const handleResetAdjustments = useCallback(
-    (paths?: Array<string>) => {
-      const pathsToReset = paths || multiSelectedPaths;
-      if (pathsToReset.length === 0) {
-        return;
-      }
-
-      debouncedSetHistory.cancel();
-
-      invoke(Invokes.ResetAdjustmentsForPaths, { paths: pathsToReset })
-        .then(() => {
-          if (libraryActivePath && pathsToReset.includes(libraryActivePath)) {
-            setLibraryActiveAdjustments((prev: Adjustments) => ({ ...INITIAL_ADJUSTMENTS, rating: prev.rating }));
-          }
-          if (selectedImage && pathsToReset.includes(selectedImage.path)) {
-            const currentRating = adjustments.rating;
-
-            const originalAspectRatio =
-              selectedImage.width && selectedImage.height ? selectedImage.width / selectedImage.height : null;
-
-            resetAdjustmentsHistory({
-              ...INITIAL_ADJUSTMENTS,
-              aspectRatio: originalAspectRatio,
-              rating: currentRating,
-              aiPatches: [],
-            });
-          }
-        })
-        .catch((err) => {
-          console.error('Failed to reset adjustments:', err);
-          setError(`Failed to reset adjustments: ${err}`);
-        });
-    },
-    [
-      multiSelectedPaths,
-      libraryActivePath,
-      selectedImage,
-      adjustments.rating,
-      resetAdjustmentsHistory,
-      debouncedSetHistory,
-    ],
-  );
-
-  const handleImportClick = useCallback(
-    async (targetPath: string) => {
-      try {
-        const nonRaw = supportedTypes?.nonRaw || [];
-        const raw = supportedTypes?.raw || [];
-
-        const expandExtensions = (exts: string[]) => {
-          return Array.from(new Set(exts.flatMap((ext) => [ext.toLowerCase(), ext.toUpperCase()])));
-        };
-
-        const processedNonRaw = expandExtensions(nonRaw);
-        const processedRaw = expandExtensions(raw);
-        const allImageExtensions = [...processedNonRaw, ...processedRaw];
-
-        const selected = await open({
-          filters: [
-            {
-              name: 'All Supported Images',
-              extensions: allImageExtensions,
-            },
-            {
-              name: 'RAW Images',
-              extensions: processedRaw,
-            },
-            {
-              name: 'Standard Images (JPEG, PNG, etc.)',
-              extensions: processedNonRaw,
-            },
-            {
-              name: 'All Files',
-              extensions: ['*'],
-            },
-          ],
-          multiple: true,
-          title: 'Select files to import',
-        });
-
-        if (Array.isArray(selected) && selected.length > 0) {
-          setImportSourcePaths(selected);
-          setImportTargetFolder(targetPath);
-          setIsImportModalOpen(true);
-        }
-      } catch (err) {
-        console.error('Failed to open file dialog for import:', err);
-      }
-    },
-    [supportedTypes],
-  );
 
   const handleEditorContextMenu = (event: any) => {
     event.preventDefault();
