+++ src/hooks/handlers/useHandleMultiSelectClick.ts
@@ -1,0 +1,57 @@
+import { MultiSelectOptions } from '../../App';
+import { ImageFile } from '../../components/ui/AppProperties';
+import { useAppState } from '../../context/ContextProviders';
+import { useSortedImageList } from '../useSortedImageList';
+
+export function useHandleMultiselectClick() {
+  const { multiSelectedPaths, setMultiSelectedPaths, setLibraryActivePath } = useAppState();
+  const { sortedImageList } = useSortedImageList();
+
+  const handleMultiSelectClick = (path: string, event: any, options: MultiSelectOptions) => {
+    const { ctrlKey, metaKey, shiftKey } = event;
+    const isCtrlPressed = ctrlKey || metaKey;
+    const { shiftAnchor, onSimpleClick, updateLibraryActivePath } = options;
+
+    if (shiftKey && shiftAnchor) {
+      const lastIndex = sortedImageList.findIndex((f) => f.path === shiftAnchor);
+      const currentIndex = sortedImageList.findIndex((f) => f.path === path);
+
+      if (lastIndex !== -1 && currentIndex !== -1) {
+        const start = Math.min(lastIndex, currentIndex);
+        const end = Math.max(lastIndex, currentIndex);
+        const range = sortedImageList.slice(start, end + 1).map((f: ImageFile) => f.path);
+        const baseSelection = isCtrlPressed ? multiSelectedPaths : [shiftAnchor];
+        const newSelection = Array.from(new Set([...baseSelection, ...range]));
+
+        setMultiSelectedPaths(newSelection);
+        if (updateLibraryActivePath) {
+          setLibraryActivePath(path);
+        }
+      }
+    } else if (isCtrlPressed) {
+      const newSelection = new Set(multiSelectedPaths);
+      if (newSelection.has(path)) {
+        newSelection.delete(path);
+      } else {
+        newSelection.add(path);
+      }
+
+      const newSelectionArray = Array.from(newSelection);
+      setMultiSelectedPaths(newSelectionArray);
+
+      if (updateLibraryActivePath) {
+        if (newSelectionArray.includes(path)) {
+          setLibraryActivePath(path);
+        } else if (newSelectionArray.length > 0) {
+          setLibraryActivePath(newSelectionArray[newSelectionArray.length - 1]);
+        } else {
+          setLibraryActivePath(null);
+        }
+      }
+    } else {
+      onSimpleClick(path);
+    }
+  };
+
+  return handleMultiSelectClick;
+}
