+++ src/hooks/handlers/useHandlePasteFiles.ts
@@ -1,0 +1,32 @@
+import { invoke } from '@tauri-apps/api/core';
+import { useCallback } from 'react';
+import { Invokes } from '../../components/ui/AppProperties';
+import { useAppState } from '../../context/ContextProviders';
+import { useRefreshImageList } from './useRefreshImageList';
+
+export function useHandlePasteFiles() {
+  const { copiedFilePaths, currentFolderPath, setCopiedFilePaths, setError } = useAppState();
+  const refreshImageList = useRefreshImageList();
+
+  const handlePasteFiles = useCallback(
+    async (mode = 'copy') => {
+      if (copiedFilePaths.length === 0 || !currentFolderPath) {
+        return;
+      }
+      try {
+        if (mode === 'copy')
+          await invoke(Invokes.CopyFiles, { sourcePaths: copiedFilePaths, destinationFolder: currentFolderPath });
+        else {
+          await invoke(Invokes.MoveFiles, { sourcePaths: copiedFilePaths, destinationFolder: currentFolderPath });
+          setCopiedFilePaths([]);
+        }
+        await refreshImageList();
+      } catch (err) {
+        setError(`Failed to ${mode} files: ${err}`);
+      }
+    },
+    [copiedFilePaths, currentFolderPath, refreshImageList],
+  );
+
+  return handlePasteFiles;
+}
